ROUGH. Tests working. Had to revert pytest. Essentially removes codewof/ folder.
