+++ codewof/codewof/admin.py
@@ -1,18 +1,18 @@
-"""Admin configuration for codeWOF."""
+# """Admin configuration for codeWOF."""
 
-from django.contrib import admin
-from django.contrib.auth import get_user_model
-from programming.models import Attempt, Badge
+# from django.contrib import admin
+# from django.contrib.auth import get_user_model
+# from programming.models import Attempt, Badge
 
-User = get_user_model()
-
-
-class AttemptAdmin(admin.ModelAdmin):
-    """Configuration for displaying attempts in admin."""
-
-    list_display = ('datetime', 'question', 'profile', 'passed_tests')
-    ordering = ('-datetime', )
+# User = get_user_model()
 
 
-admin.site.register(Attempt, AttemptAdmin)
-admin.site.register(Badge)
+# class AttemptAdmin(admin.ModelAdmin):
+#     """Configuration for displaying attempts in admin."""
+
+#     list_display = ('datetime', 'question', 'profile', 'passed_tests')
+#     ordering = ('-datetime', )
+
+
+# admin.site.register(Attempt, AttemptAdmin)
+# admin.site.register(Badge)
