+++ codewof/tests/codewof_test_data_generator.py
@@ -1,9 +1,13 @@
 """Class to generate test data required for testing codewof system."""
 
 from django.contrib.auth import get_user_model
+from django.core import management
 from datetime import datetime
 
 from programming.models import Badge, Question, Attempt
+
+from tests.users.factories import UserFactory
+from users.models import UserType
 
 User = get_user_model()
 
@@ -15,10 +19,9 @@ def generate_questions():
 
 def generate_users(user):
     """Generate users for codeWOF tests. Creates two basic users for unit tests."""
-    # call_command("load_user_types")
-    # user.create_batch(size=2)
-    # User.objects.create_user(username='john', email='john@uclive.ac.nz', password='onion', user_type=usertype)
-    # User.objects.create_user(username='sally', email='sally@uclive.ac.nz', password='onion', user_type=usertype)
+    management.call_command("load_user_types")
+    User.objects.create_user(id=1, username='john', email='john@uclive.ac.nz', password='onion', user_type=UserType.objects.get(slug='student'))
+    User.objects.create_user(id=2, username='sally', email='sally@uclive.ac.nz', password='onion', user_type=UserType.objects.get(slug='other'))
 
 
 def generate_badges():
