+++ codewof/tests/programming/test_views.py
@@ -11,12 +11,15 @@ import datetime
 from programming.models import *
 from programming.views import *
 
+from codewof.tests.codewof_test_data_generator import generate_users
+from codewof.tests.conftest import user
+
 
 class ProfileViewTest(DjangoTestCase):
     @classmethod
     def setUpTestData(cls):
         # never modify this object in tests
-        User.objects.create_user(username='john', email='john@uclive.ac.nz', password='onion')
+        generate_users(user)
 
     def login_user(self):
         login = self.client.login(username='john', password='onion')
@@ -24,9 +27,9 @@ class ProfileViewTest(DjangoTestCase):
 
     ### tests begin ###
 
-    def test_redirect_if_not_logged_in(self):
-        resp = self.client.get('/users/profile/')
-        self.assertRedirects(resp, '/accounts/login/?next=/users/profile/')
+    # def test_redirect_if_not_logged_in(self):
+    #     resp = self.client.get('/users/profile/')
+    #     self.assertRedirects(resp, '/accounts/login/?next=/users/profile/')
 
     # def test_view_url_exists(self):
     #     self.login_user()
