+++ requirements/local.txt
@@ -6,7 +6,7 @@ psycopg2-binary==2.8.4  # https://github.com/psycopg/psycopg2
 # Testing
 # ------------------------------------------------------------------------------
 mypy==0.770  # https://github.com/python/mypy
-pytest==5.4.1  # https://github.com/pytest-dev/pytest
+pytest==5.3.5  # https://github.com/pytest-dev/pytest
 pytest-sugar==0.9.2  # https://github.com/Frozenball/pytest-sugar
 
 # Code quality
